refactor: merge EditorInstance into TrimeInputMethodService
EditorInstance was created in the past for preparing to convert TrimeInputMethodService source to Kotlin, since all our source code are in Kotlin now, it seems to be redundant, so I terminate it in this commit.
